requirements.txt
beautifulsoup4==4.10.0
coverage==6.2
pytest==7.0.1
pyyaml==6.0
selenium==4.1.0
webdriver-manager==3.8.3